# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "App-demo Documentation",
    "version": "1.0.0",
    "title": "App-demo"
  },
  "paths": {
...
    "/echo": {
      "post": {
        "tags": [
          "features"
        ],
        "summary": "Write in logs",
        "responses": {
          "204": {
            "description": "Successful"
          },
          "500": {
            "description": "App health's is bad"
          }
        }
      }
    },
    "/env": {
      "get": {
        "tags": [
          "features"
        ],
        "summary": "Get enviroment variables",
        "responses": {
          "200": {
            "description": "successful",
            "schema": {
              "type": "object",
              "properties": {
                "env-1": {
                  "type": "string"
                },
                "env-2": {
                  "type": "string"
                },
                "env-3": {
                  "type": "string"
                },
                "env-4": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "App health's is bad"
          }
        }
      }
    },
    "/headers": {
      "get": {
        "tags": [
          "features"
        ],
        "summary": "Get requests headers",
        "responses": {
          "200": {
            "description": "successful",
            "schema": {
              "type": "object",
              "properties": {
                "env-1": {
                  "type": "string"
                },
                "env-2": {
                  "type": "string"
                },
                "env-3": {
                  "type": "string"
                },
                "env-4": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "App health's is bad"
          }
        }
      }
    }
  }
}